requirements.txt
streamlit==1.38.0
PyPDF2==3.0.1
python-docx==1.1.2
pdf2image==1.17.0
pytesseract==0.3.13
spacy==3.7.6